Remove state table.
Retrieve new comments based on last execution. Wanted to avoid doing it this way, but this is more than adequate. Comments may be missed if the function is paused or errors out (e.g. Reddit API errors), but that hasn't been an issue thus far. If the state table is needed for something else in the future, this may be worth revisiting. Changed ThreadTimeout to seconds for consistency. Updated version.